chore(lane_change): move common utils
