+++ planning/behavior_path_planner/include/behavior_path_planner/util/lane_change/debug.hpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__LANE_CHANGE__DEBUG_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__LANE_CHANGE__DEBUG_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__LANE_CHANGE__DEBUG_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__LANE_CHANGE__DEBUG_HPP_
 
 #include "behavior_path_planner/debug_utilities.hpp"
-#include "behavior_path_planner/scene_module/lane_change/lane_change_path.hpp"
+#include "behavior_path_planner/util/lane_change/lane_change_path.hpp"
 
 #include <string>
 #include <unordered_map>
@@ -40,4 +40,4 @@ MarkerArray showPolygon(
 MarkerArray showPolygonPose(
   const std::unordered_map<std::string, CollisionCheckDebug> & obj_debug_vec, std::string && ns);
 }  // namespace marker_utils::lane_change_markers
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__LANE_CHANGE__DEBUG_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__LANE_CHANGE__DEBUG_HPP_
